Face_Landmarks/requirements.txt
absl-py==2.2.1
altgraph==0.17.4
aniso8601==10.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.37.28
botocore==1.37.28
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
flatbuffers==25.2.10
fonttools==4.56.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
jmespath==1.0.1
jsons==1.6.3
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
s3transfer==0.11.4
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.1
typish==1.9.3
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
